EXAMS/exam P10 Toirjonov Odiljon M6/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-environ==0.10.0
environ==1.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
psycopg2==2.9.6
Pillow==9.5.0
django-filter~=23.2
facebook-sdk==3.1.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google~=3.0.0
mongoengine~=0.27.0